+++ seantis/dir/events/tests/test_browser.py
@@ -194,13 +194,13 @@ class CommonBrowserTests(BrowserTestCase):
 
         # regression test for an issue where occurrences in the future
         # were not counted correctly
-        fourchan.set_date('submission_date', datetime(2021, 1, 1, 0, 0))
+        fourchan.set_date('submission_date', datetime(2019, 1, 1, 0, 0))
         fourchan.widget('submission_start_time').value = '10:00 AM'
         fourchan.widget('submission_end_time').value = '11:00 AM'
 
         fourchan.widget(
             'submission_recurrence'
-        ).value = 'RRULE:FREQ=DAILY;UNTIL=20211231T000000'
+        ).value = 'RRULE:FREQ=DAILY;UNTIL=20191231T000000'
 
         fourchan.getControl('Continue').click()  # ok
 
@@ -212,7 +212,7 @@ class CommonBrowserTests(BrowserTestCase):
 
         fourchan.widget(
             'submission_recurrence'
-        ).value = 'RRULE:FREQ=DAILY;UNTIL=20220101T000000'
+        ).value = 'RRULE:FREQ=DAILY;UNTIL=20200101T000000'
 
         fourchan.getControl('Continue').click()  # not okay